Module Lambdasoup.R

For each function Soup.f that evaluates to an option, Soup.R.f is a version of f that is post-composed with require, so, for example, you can write soup |> children |> R.first instead of soup |> children |> first |> require.

val select_one : string -> _ node -> element node

val attribute : string -> element node -> string

val id : element node -> string

val element : _ node -> element node

val leaf_text : _ node -> string

val nth : int -> 'a nodes -> 'a node

val first : 'a nodes -> 'a node

val last : 'a nodes -> 'a node

val tag : string -> _ node -> element node

val parent : _ node -> element node

val child : _ node -> general node

val child_element : _ node -> element node

val next_sibling : _ node -> general node

val previous_sibling : _ node -> general node

val next_element : _ node -> element node

val previous_element : _ node -> element node